+++ src/main/java/com/azer467/saphyre/items/GiftItem.java
@@ -1,14 +1,11 @@
-package com.azer467.saphyre.objects.items;
+package com.azer467.saphyre.items;
 
 import com.azer467.saphyre.SaphyreMetadata;
 import net.minecraft.Util;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.resources.language.I18n;
+import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
@@ -19,9 +16,8 @@ import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.LootTable;
-import net.minecraft.world.level.storage.loot.LootTables;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSet;
-import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.minecraftforge.registries.RegistryObject;
 
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
@@ -29,13 +25,21 @@ import java.util.concurrent.ThreadLocalRandom;
 public class GiftItem extends Item {
     private final String itemName;
     //TODO: custom loot tables ?
-    private String lootTable;
+    private final String lootTable;
+
+    // We use the registry object entry instead of directly calling the object since it's not yet registered into the particle registry
+    private RegistryObject<SimpleParticleType> giftParticle;
 
     public GiftItem(String itemName, String lootTable, Properties properties)
     {
         super(properties);
         this.itemName = itemName;
         this.lootTable = lootTable;
+    }
+
+    public GiftItem setGiftParticle(RegistryObject<SimpleParticleType> particleType) {
+        this.giftParticle = particleType;
+        return this;
     }
 
     @Override
@@ -51,6 +55,7 @@ public class GiftItem extends Item {
         if (level.isClientSide) {
             ItemStack currentSaphyre = player.getItemInHand(interactionHand);
             player.startUsingItem(interactionHand);
+            spawnGiftParticles(level, player);
             return InteractionResultHolder.success(currentSaphyre);
         } else {
             LootTable giftTable = ((ServerLevel) level).getServer().getLootTables().get(new ResourceLocation(lootTable));
@@ -68,17 +73,23 @@ public class GiftItem extends Item {
         }
     }
 
+    private void spawnGiftParticles(Level level, Player player) {
+        for (int i = 0; i < 360; i++) {
+            if (i % 20 == 0) {
+                level.addParticle(
+                    giftParticle.get(),
+                    player.getX(),
+                    player.getY() + 0.50,
+                    player.getZ(),
+                    Math.cos(i) * 0.25d,
+                    0.15d,
+                    Math.sin(i) * 0.25d
+                );
+            }
+        }
+    }
+
     private TranslatableComponent formatSuccessChatMessage(ItemStack itemStack) {
         return new TranslatableComponent("message.gifting.item.success", itemStack.getCount(), itemStack.getDisplayName());
     }
-
-    //    public @NotNull ActionResult<ItemStack> onItemRightClick(@NotNull World worldIn, @NotNull EntityPlayer playerIn, @NotNull EnumHand handIn) {
-//        IGiftBase randomItem = dropList.sample();
-//        boolean handleSuccess = randomItem.executeHandler(worldIn, playerIn, handIn);
-//        if (handleSuccess) {
-//            return new ActionResult<>(EnumActionResult.PASS, playerIn.getHeldItem(handIn));
-//        } else {
-//            return new ActionResult<>(EnumActionResult.FAIL, playerIn.getHeldItem(handIn));
-//        }
-//    }
 }