+++ CHANGELOG.md
@@ -20,6 +20,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Code Cleanup
 - Folder Structure
 
+### Fixed
+
+- Dependencies constraints
+
 ## [0.3.0-118-FORGE] - 06/06/2022
 
 ### Added