+++ CHANGELOG.md
@@ -8,7 +8,17 @@ When a version is ported to a new game release, it starts from the preceding ver
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [0.4.0-118-FORGE] - UNRELEASED
+## [0.5.0-118-FORGE] - UNRELEASED
+
+### Added 
+
+- Particles
+
+### Changed
+
+- Moved Particle & Entity inits to init package
+
+## [0.4.0-118-FORGE] - 19/06/2022
 
 ### Added